wang-6

+++ src/main/java/com/geek/geekstudio/service/proxy/FileServiceProxy.java
@@ -6,6 +6,7 @@ import com.geek.geekstudio.exception.RecruitFileException;
 import com.geek.geekstudio.mapper.*;
 import com.geek.geekstudio.model.po.TaskPO;
 import com.geek.geekstudio.model.po.WorkPO;
+import com.geek.geekstudio.model.vo.ErrorMsg;
 import com.geek.geekstudio.model.vo.RestInfo;
 import com.geek.geekstudio.service.FileService;
 import com.geek.geekstudio.service.impl.FileServiceImpl;
@@ -16,6 +17,8 @@ import org.springframework.stereotype.Service;
 import org.springframework.web.multipart.MultipartFile;
 
 import java.io.File;
+import java.util.ArrayList;
+import java.util.List;
 
 @Service
 @Slf4j
@@ -85,13 +88,15 @@ public class FileServiceProxy implements FileService {
         try {
             return fileService.taskFileUpload(taskId, file);
         } catch (RecruitFileException e) {
-            //在标有@Transcation的方法上不能通过try-catch去维护数据一致性
+            //在标有@Transcation的方法上不能通过try-catch去维护数据库上的数据一致性
             //因为一切操作数据库语句都会回滚，但是操作到磁盘上的操作不会回滚(所以可能会有文件残留)
             //回滚（一致性）策略，若该任务没有对应的一个文件，就删除task表上的记录
             if (taskFileMapper.queryFileCount(taskId) == 0) {
                 //taskMapper.deleteByTaskId(taskId);
-                TaskPO taskPO=taskMapper.queryOneTaskById(taskId);
-                fileUtil.deleteDir(new File(fileUtil.buildTaskFilePath(taskPO.getCourseId(),taskId)));
+                TaskPO taskPO = taskMapper.queryOneTaskById(taskId);
+                if (taskPO != null) {
+                    fileUtil.deleteDir(new File(fileUtil.buildTaskFilePath(taskPO.getCourseId(), taskId)));
+                }
             }
             throw new RecruitFileException(e.getMessage());
         }
@@ -103,7 +108,16 @@ public class FileServiceProxy implements FileService {
     @Override
     public RestInfo taskFilesUpload(int taskId, MultipartFile[] files) {
         log.info("任务编号为" + taskId + " 上传了多个文件");
-        return fileService.taskFilesUpload(taskId, files);
+        int total = files.length;
+        List<ErrorMsg> errorList = new ArrayList<>();
+        for (int i = 0; i < total; i++) {
+            try {
+                this.taskFileUpload(taskId, files[i]);
+            } catch (RecruitFileException e) {
+                errorList.add(new ErrorMsg(405, "第" + (i + 1) + "个文件上传失败!", e.getMessage()));
+            }
+        }
+        return RestInfo.success("要布置" + total + "个作业，" + "成功布置" + (total - errorList.size()) + "个作业文件", errorList);
     }
 
     /**
@@ -116,7 +130,7 @@ public class FileServiceProxy implements FileService {
     }
 
     /**
-     *上交作业的文件上传
+     * 上交作业的文件上传
      */
     @Override
     public RestInfo workFileUpload(int workId, MultipartFile file) throws RecruitFileException {
@@ -126,28 +140,41 @@ public class FileServiceProxy implements FileService {
         } catch (RecruitFileException e) {
             //回滚（一致性）策略，若该作业记录没有对应的一个文件，就删除work表上的记录
             if (workFileMapper.queryFileCount(workId) == 0) {
-                WorkPO workPO=workMapper.queryWorkById(workId);
-                workMapper.deleteByWorkId(workId);
-                fileUtil.deleteDir(new File(fileUtil.buildWorkFilePath(workPO.getCourseId(),workPO.getUserId(),workPO.getTaskId())));
+                WorkPO workPO = workMapper.queryWorkById(workId);
+                if (workPO != null) {
+                    workMapper.deleteByWorkId(workId);
+                    fileUtil.deleteDir(new File(fileUtil.buildWorkFilePath(workPO.getCourseId(), workPO.getUserId(), workPO.getTaskId())));
+                }
             }
             throw new RecruitFileException(e.getMessage());
         }
     }
 
     /**
-     *多个作业的文件上传
+     * 多个作业的文件上传
      */
     @Override
     public RestInfo workFilesUpload(int workId, MultipartFile[] files) {
         log.info("任务编号为" + workId + " 上传了多个文件");
-        RestInfo restInfo=fileService.workFilesUpload(workId, files);
+        int total = files.length;
+        List<ErrorMsg> errorList = new ArrayList<>();
+        for (int i = 0; i < total; i++) {
+            try {
+                //同一个类内调用@Transactional修饰的方法，事务不生效
+                fileService.workFileUpload(workId, files[i]);
+            } catch (RecruitFileException e) {
+                errorList.add(new ErrorMsg(405, "第" + (i + 1) + "个作业文件上传失败!", e.getMessage()));
+            }
+        }
         //如果一个文件都没有上传成功就删除原work记录，清理磁盘上可能存在的未上传成功文件碎片
         if (workFileMapper.queryFileCount(workId) == 0) {
-            WorkPO workPO=workMapper.queryWorkById(workId);
-            workMapper.deleteByWorkId(workId);
-            fileUtil.deleteDir(new File(fileUtil.buildWorkFilePath(workPO.getCourseId(),workPO.getUserId(),workPO.getTaskId())));
+            WorkPO workPO = workMapper.queryWorkById(workId);
+            if (workPO != null) {
+                workMapper.deleteByWorkId(workId);
+                fileUtil.deleteDir(new File(fileUtil.buildWorkFilePath(workPO.getCourseId(), workPO.getUserId(), workPO.getTaskId())));
+            }
         }
-        return restInfo;
+        return RestInfo.success("成功提交" + (total - errorList.size()) + "个作业文件", errorList);
     }
 
     /**
@@ -156,6 +183,6 @@ public class FileServiceProxy implements FileService {
     @Override
     public RestInfo delWorkFile(int id, String userId) throws RecruitException {
         log.info("用户编号为：" + userId + " 尝试删除已上传作业编号为" + id + "的文件");
-        return fileService.delWorkFile(id,userId);
+        return fileService.delWorkFile(id, userId);
     }
 }
